+++ lib/spack/spack/operating_systems/cray_frontend.py
@@ -11,9 +11,10 @@ import llnl.util.filesystem as fs
 import llnl.util.lang
 import llnl.util.tty as tty
 
-from spack.operating_systems.linux_distro import LinuxDistro
 from spack.util.environment import get_path
 from spack.util.module_cmd import module
+
+from .linux_distro import LinuxDistro
 
 
 @contextlib.contextmanager
