+++ lib/spack/spack/operating_systems/linux_distro.py
@@ -2,10 +2,9 @@
 # Spack Project Developers. See the top-level COPYRIGHT file for details.
 #
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
-
 import re
 
-from spack.architecture import OperatingSystem
+from ._base_classes import OperatingSystem
 
 
 class LinuxDistro(OperatingSystem):
