+++ lib/spack/spack/platforms/darwin.py
@@ -7,8 +7,10 @@ import platform
 
 import archspec.cpu
 
-from spack.architecture import Platform, Target
+import spack.target
 from spack.operating_systems.mac_os import MacOs
+
+from ._base_classes import Platform
 
 
 class Darwin(Platform):
@@ -20,7 +22,7 @@ class Darwin(Platform):
         super(Darwin, self).__init__('darwin')
 
         for name in archspec.cpu.TARGETS:
-            self.add_target(name, Target(name))
+            self.add_target(name, spack.target.Target(name))
 
         self.default = archspec.cpu.host().name
         self.front_end = self.default
