+++ lib/spack/spack/platforms/linux.py
@@ -6,8 +6,10 @@ import platform
 
 import archspec.cpu
 
-from spack.architecture import Platform, Target
+import spack.target
 from spack.operating_systems.linux_distro import LinuxDistro
+
+from ._base_classes import Platform
 
 
 class Linux(Platform):
@@ -17,7 +19,7 @@ class Linux(Platform):
         super(Linux, self).__init__('linux')
 
         for name in archspec.cpu.TARGETS:
-            self.add_target(name, Target(name))
+            self.add_target(name, spack.target.Target(name))
 
         # Get specific default
         self.default = archspec.cpu.host().name
