+++ lib/spack/spack/platforms/test.py
@@ -2,10 +2,12 @@
 # Spack Project Developers. See the top-level COPYRIGHT file for details.
 #
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
-
 import platform
 
-from spack.architecture import OperatingSystem, Platform, Target
+import spack.operating_systems
+import spack.target
+
+from ._base_classes import Platform
 
 
 class Test(Platform):
@@ -24,13 +26,13 @@ class Test(Platform):
 
     def __init__(self):
         super(Test, self).__init__('test')
-        self.add_target(self.default, Target(self.default))
-        self.add_target(self.front_end, Target(self.front_end))
+        self.add_target(self.default, spack.target.Target(self.default))
+        self.add_target(self.front_end, spack.target.Target(self.front_end))
 
         self.add_operating_system(
-            self.default_os, OperatingSystem('debian', 6))
+            self.default_os, spack.operating_systems.OperatingSystem('debian', 6))
         self.add_operating_system(
-            self.front_os, OperatingSystem('redhat', 6))
+            self.front_os, spack.operating_systems.OperatingSystem('redhat', 6))
 
     @classmethod
     def detect(cls):
